+++ src/ll/xist/ns/form.py
@@ -9,7 +9,7 @@
 
 
 """
-An XIST module that contains elements that are useful for forms. These are
+This XIST module contains convenience classes for form elements. These are
 just abbreviations for the various ``<input type="...">`` elements.
 """
 
@@ -24,6 +24,16 @@ __docformat__ = "reStructuredText"
 xmlns = "http://xmlns.livinglogic.de/xist/ns/form"
 
 
+class text(html.input):
+	xmlns = xmlns
+	class Attrs(html.input.Attrs):
+		type = None
+
+	def convert(self, converter):
+		e = html.input(self.attrs, type=u"text")
+		return e.convert(converter)
+
+
 class checkbox(html.input):
 	xmlns = xmlns
 	class Attrs(html.input.Attrs):
@@ -31,16 +41,6 @@ class checkbox(html.input):
 
 	def convert(self, converter):
 		e = html.input(self.attrs, type=u"checkbox")
-		return e.convert(converter)
-
-
-class edit(html.input):
-	xmlns = xmlns
-	class Attrs(html.input.Attrs):
-		type = None
-
-	def convert(self, converter):
-		e = html.input(self.attrs, type=u"text")
 		return e.convert(converter)
 
 
@@ -54,6 +54,16 @@ class radio(html.input):
 		return e.convert(converter)
 
 
+class file(html.input):
+	xmlns = xmlns
+	class Attrs(html.input.Attrs):
+		type = None
+
+	def convert(self, converter):
+		e = html.input(self.attrs, type=u"file")
+		return e.convert(converter)
+
+
 class submit(html.input):
 	xmlns = xmlns
 	class Attrs(html.input.Attrs):
@@ -64,13 +74,23 @@ class submit(html.input):
 		return e.convert(converter)
 
 
-class memo(html.textarea):
+class button(html.input):
+	xmlns = xmlns
+	class Attrs(html.input.Attrs):
+		type = None
+
+	def convert(self, converter):
+		e = html.input(self.attrs, type=u"button")
+		return e.convert(converter)
+
+
+class textarea(html.textarea):
 	xmlns = xmlns
 	class Attrs(html.textarea.Attrs):
 		class value(xsc.TextAttr): pass
 
 	def convert(self, converter):
-		e = html.textarea(self[u"value"], self.attrs.withoutnames(u"value"))
+		e = html.textarea(self.attrs.value, self.attrs.withoutnames(u"value"))
 		return e.convert(converter)